+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added `endAnchor` parameter to dictate the window of interest.
 * Added `windowEnd` parameter to dictate the window of interest.
 * Added `concatTargets` parameter to specify if multiple target cohorts should be treated separately, or continuously.
+* Added `overlapMethod` parameter to pick a method how to deal with non-combination overlap between events.
 * Removed `indexOffset` parameter.
 * Removed `includeTreatments` parameter.